+++ lib/index.d.ts
@@ -1,0 +1,2 @@
+export { setConfig, wpSiteInfo, wpPostBySlug, wpPublishedPostsSlugsByCategoryId, wpLatestPublishedPostsByCategoryId } from "./functions";
+export { formatDate, getTextStringFromHtml } from "./support";