+++ lib/support.d.ts
@@ -1,0 +1,2 @@
+export declare const getTextStringFromHtml: (html: string, wordCount?: number) => string;
+export declare const formatDate: (date: string, locale?: string) => string;