+++ lib/types.d.ts
@@ -1,0 +1,4 @@
+export type WpSite = {
+    url: string | undefined;
+    wpGraphqlUrl: string | undefined;
+};